# PoC/12-housing-board/schemas/api/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "주택파동 지도 API",
  "version": "1.0.0",
  "description": "번들이 없는 지역도 지도가 열리도록 하는 요청 시 조회 API. 열리는 방식은 3등급: A 번들(regions/<slug>) · B 번들 없는 법정동·필지(경계·건물·인허가·기반시설 요청 시 조회) · C 번들 없는 시군구(경계만). 이 문서가 서버·화면 사이의 계약이며 tests/js/contract.test.cjs 가 실제 응답이 이 스키마를 지키는지 확인한다.",
  "license": {
   "name": "UNLICENSED"
  }
 },
 "servers": [
  {
   "url": "https://housing-board.vercel.app",
   "description": "운영"
  },
  {
   "url": "http://localhost:8000",
   "description": "로컬 개발 서버(node scripts/dev.js 8000)"
  }
 ],
...
 "paths": {
...
  "/api/v1/permits": {
   "get": {
    "operationId": "getPermits",
    "summary": "인허가 사업 요청 시 조회",
    "description": "법정동 하나의 건축HUB 주택인허가를 번지(PNU) 단위 사업 + 필지 경계로 돌려준다. 공공주택지구 블록 단위 허가의 위치·합필 지번·준공은 건축물대장 총괄표제부로 보강한다.",
    "tags": [
     "permits"
    ],
    "parameters": [
     {
      "name": "bjd",
      "in": "query",
      "required": true,
      "description": "법정동 8자리(00 보충) 또는 10자리",
      "schema": {
       "type": "string",
       "pattern": "^\\d{8}(\\d{2})?$"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "사업 목록(한 법정동 최대 120곳). 필지를 일부 못 받았거나 건물대장 보강이 완전하지 않으면 s-maxage=60",
      "headers": {
       "Cache-Control": {
        "description": "public, s-maxage=86400, stale-while-revalidate=86400",
        "schema": {
         "type": "string"
        }
       }
      },
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/PermitsResponse"
        }
       }
      }
     },
     "400": {
      "description": "invalid-query · invalid-code(법정동이 아님)",
      "content": {
       "application/problem+json": {
        "schema": {
         "$ref": "#/components/schemas/Problem"
        }
       }
      }
     },
     "429": {
      "description": "keys-exhausted(키 한도) 또는 budget-exhausted(시간당 필지 호출 상한). Retry-After",
      "content": {
       "application/problem+json": {
        "schema": {
         "$ref": "#/components/schemas/Problem"
        }
       }
      }
     },
     "405": {
      "description": "GET·HEAD 만 허용(Allow 헤더)",
      "content": {
       "application/problem+json": {
        "schema": {
         "$ref": "#/components/schemas/Problem"
        }
       }
      }
     },
     "502": {
      "description": "원천 서비스 오류(원인 문구는 싣지 않음). Cache-Control: no-store",
      "content": {
       "application/problem+json": {
        "schema": {
         "$ref": "#/components/schemas/Problem"
        }
       }
      }
     },
     "503": {
      "description": "인증키가 없음. Cache-Control: no-store",
      "content": {
       "application/problem+json": {
        "schema": {
         "$ref": "#/components/schemas/Problem"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/infra": {
   "get": {
    "operationId": "getInfra",
    "summary": "기반시설(신설예정 학교·버스 정류장) 요청 시 조회",
    "description": "법정동 인허가 단지 가까이의 신설예정 학교(교육재정알리미)와 버스 정류장(TAGO, 없으면 OpenStreetMap)을 번들 infra.json 과 같은 모양으로 돌려준다. 서버는 임의 좌표를 받지 않고 그 법정동 인허가 결과의 필지 중심만 쓴다. 인허가는 이 서버의 permits API 를 공개 주소로 불러 CDN 캐시에서 받는다.",
    "tags": [
     "infra"
    ],
    "parameters": [
     {
      "name": "bjd",
      "in": "query",
      "required": true,
      "description": "법정동 8자리(00 보충) 또는 10자리",
      "schema": {
       "type": "string",
       "pattern": "^\\d{8}(\\d{2})?$"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "학교·정류장. 한쪽이 실패하면 meta.schoolsError·stopsError 와 s-maxage=60. TAGO 밖 지역: 서울(시도 11)은 서울특별시 정류소정보조회(meta.stopsSource='seoul'), 실패하면 OpenStreetMap(meta.stopsSource='osm', meta.seoulError), 강릉 등 그 밖은 OpenStreetMap, 그것도 못 받으면 meta.noBus",
      "headers": {
       "Cache-Control": {
        "description": "public, s-maxage=86400, stale-while-revalidate=86400",
        "schema": {
         "type": "string"
        }
       }
      },
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InfraResponse"
        }
       }
      }
     },
     "400": {
      "description": "invalid-query · invalid-code",
      "content": {
       "application/problem+json": {
        "schema": {
         "$ref": "#/components/schemas/Problem"
        }
       }
      }
     },
     "429": {
      "description": "keys-exhausted 또는 budget-exhausted(시간당 TAGO 호출 상한). Retry-After",
      "content": {
       "application/problem+json": {
        "schema": {
         "$ref": "#/components/schemas/Problem"
        }
       }
      }
     },
     "405": {
      "description": "GET·HEAD 만 허용(Allow 헤더)",
      "content": {
       "application/problem+json": {
        "schema": {
         "$ref": "#/components/schemas/Problem"
        }
       }
      }
     },
     "502": {
      "description": "원천 서비스 오류(원인 문구는 싣지 않음). Cache-Control: no-store",
      "content": {
       "application/problem+json": {
        "schema": {
         "$ref": "#/components/schemas/Problem"
        }
       }
      }
     },
     "503": {
      "description": "인증키가 없음. Cache-Control: no-store",
      "content": {
       "application/problem+json": {
        "schema": {
         "$ref": "#/components/schemas/Problem"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "Problem": {
    "type": "object",
    "description": "RFC 7807 problem+json. type 은 '/problems/<code>' 이고 code 는 아래 목록 중 하나. 오류 문구에는 인증키·원천 URL·원천 오류 문구를 싣지 않는다.",
    "required": [
     "type",
     "title",
     "status",
     "code",
     "detail"
    ],
    "properties": {
     "type": {
      "type": "string",
      "pattern": "^/problems/[a-z0-9-]+$"
     },
     "title": {
      "type": "string"
     },
     "status": {
      "type": "integer",
      "minimum": 400,
      "maximum": 599
     },
     "code": {
      "type": "string",
      "enum": [
       "invalid-query",
       "invalid-code",
       "invalid-cell",
       "unsupported-level",
       "unknown-code",
       "unknown-project",
       "method",
       "keys-exhausted",
       "budget-exhausted",
       "not-configured",
       "upstream",
       "out-of-range",
       "not-found"
      ]
     },
     "detail": {
      "type": "string"
     },
     "reason": {
      "type": "string",
      "description": "invalid-code·invalid-query 의 세부 사유(empty·not-digits·bad-length·unknown-sido·bad-pnu·bad-project·type-mismatch·too-short·too-long)"
     },
     "retryAfterSec": {
      "type": "integer",
      "minimum": 1,
      "description": "429 일 때 다시 시도할 때까지 초(Retry-After 헤더와 같음)"
     },
     "sgg": {
      "type": "string",
      "pattern": "^\\d{5}$"
     },
     "suggestions": {
      "type": "array",
      "items": {
       "type": "object",
       "required": [
        "bjd",
        "name"
       ],
       "properties": {
        "bjd": {
         "type": "string",
         "pattern": "^\\d{10}$"
        },
        "name": {
         "type": "string"
        }
       }
      },
      "description": "unknown-code(법정동)일 때 같은 시군구의 법정동 후보 8개 이내"
     },
     "project": {
      "type": "string",
      "pattern": "^PRJ-\\d{5}-\\d{4}$",
      "description": "unknown-project 일 때 요청한 사업 id"
     }
    },
    "additionalProperties": true
   },
...
   "PartialDate": {
    "type": "string",
    "pattern": "^\\d{4}(-\\d{2}(-\\d{2})?)?$",
    "description": "원천이 연·월까지만 적은 날짜는 YYYY-MM · YYYY 로 온다(예정일)"
   },
...
   "Geometry": {
    "type": "object",
    "description": "GeoJSON Polygon 또는 MultiPolygon(경계는 점 수를 줄여 단순화함)",
    "required": [
     "type",
     "coordinates"
    ],
    "properties": {
     "type": {
      "type": "string",
      "enum": [
       "Polygon",
       "MultiPolygon"
      ]
     },
     "coordinates": {
      "type": "array",
      "minItems": 1
     }
    }
   },
...
   "PermitProject": {
    "type": "object",
    "description": "번지(PNU) 단위로 모은 인허가 사업 하나(화면은 단지 하나로 그린다)",
    "required": [
     "pnu",
     "jibun",
     "name",
     "label",
     "units",
     "status",
     "records"
    ],
    "properties": {
     "pnu": {
      "type": "string",
      "pattern": "^\\d{19}$"
     },
     "jibun": {
      "type": "string"
     },
     "name": {
      "type": "string"
     },
     "label": {
      "type": "string",
      "description": "짧은 이름(8자 이내)"
     },
     "units": {
      "oneOf": [
       {
        "type": "integer",
        "minimum": 0
       },
       {
        "type": "null"
       }
      ],
      "description": "가장 최근 허가의 총세대수"
     },
     "status": {
      "type": "string",
      "enum": [
       "계획",
       "건설 단계",
       "입주 단계"
      ],
      "description": "인허가로 아는 3가지뿐('계획'은 착공·사용검사 기록이 원천에 없다는 뜻)"
     },
     "mainBldCnt": {
      "oneOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ]
     },
     "approvedAt": {
      "oneOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ]
     },
     "startedAt": {
      "oneOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ]
     },
     "completedAt": {
      "oneOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ]
     },
     "plannedStart": {
      "oneOf": [
       {
        "$ref": "#/components/schemas/PartialDate"
       },
       {
        "type": "null"
       }
      ]
     },
     "plannedCompletion": {
      "oneOf": [
       {
        "$ref": "#/components/schemas/PartialDate"
       },
       {
        "type": "null"
       }
      ],
      "description": "사용검사 예정일(입주 단계가 아닐 때)"
     },
     "overdue": {
      "oneOf": [
       {
        "type": "object",
        "required": [
         "kind",
         "plannedAt",
         "months"
        ],
        "properties": {
         "kind": {
          "type": "string",
          "enum": [
           "착공",
           "준공"
          ]
         },
         "plannedAt": {
          "$ref": "#/components/schemas/PartialDate"
         },
         "months": {
          "type": "integer",
          "minimum": 0
         }
        }
       },
       {
        "type": "null"
       }
      ],
      "description": "예정일이 지났는데 착공·사용검사 기록이 없다는 사실(지연 판정이 아님)"
     },
     "address": {
      "oneOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ]
     },
     "refs": {
      "type": "array",
      "items": {
       "type": "string"
      },
      "description": "그 번지에 모인 허가 관리번호(mgmHsrgstPk). 22자리까지 있어 문자열이고 숫자로 바꾸면 안 된다(건축HUB 는 JSON 숫자로 주므로 서버가 파싱 전에 문자열로 지킨다)"
     },
     "records": {
      "type": "integer",
      "minimum": 1
     },
     "latestRef": {
      "oneOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "description": "가장 최근 허가의 관리번호(문자열)"
     },
     "via": {
      "const": "ledger",
      "description": "위치를 건축물대장이 정함(블록 단위 허가 또는 합필·분할)"
     },
     "block": {
      "type": "string",
      "description": "허가가 블록 단위일 때 블록 이름"
     },
     "hubJibun": {
      "type": "string",
      "description": "합필·분할 전 허가 지번"
     },
     "statusBy": {
      "const": "ledger",
      "description": "준공을 대장 사용승인일로 올림"
     },
     "ledger": {
      "type": "object",
      "properties": {
       "name": {
        "type": "string"
       },
       "platPlc": {
        "type": "string"
       },
       "useAprDay": {
        "oneOf": [
         {
          "type": "string",
          "format": "date"
         },
         {
          "type": "null"
         }
        ]
       },
       "areaDiff": {
        "oneOf": [
         {
          "type": "number"
         },
         {
          "type": "null"
         }
        ]
       },
       "by": {
        "const": "name"
       }
      }
     },
     "projectId": {
      "type": "string",
      "pattern": "^PRJ-\\d{5}-\\d{4}$",
      "description": "사업 레지스트리에 발급된 사업이면 그 id(허가 관리번호나 PNU 가 같은 사업). 없으면 발급 전. 화면은 이 id 로 `?project=` 링크를 만든다"
     }
    }
   },
   "PermitsResponse": {
    "type": "object",
    "required": [
     "type",
     "bjd",
     "features",
     "meta"
    ],
    "properties": {
     "type": {
      "const": "FeatureCollection"
     },
     "bjd": {
      "type": "string",
      "pattern": "^\\d{10}$"
     },
     "features": {
      "type": "array",
      "maxItems": 240,
      "items": {
       "type": "object",
       "required": [
        "properties",
        "geometry"
       ],
       "properties": {
        "type": {
         "const": "Feature"
        },
        "properties": {
         "$ref": "#/components/schemas/PermitProject"
        },
        "geometry": {
         "$ref": "#/components/schemas/Geometry"
        }
       }
      }
     },
     "meta": {
      "type": "object",
      "required": [
       "source",
       "fetchedAt",
       "records",
       "pages",
       "candidates",
       "projects",
       "located",
       "unlocated",
       "blockProjects",
       "parcelErrors",
       "truncated",
       "skipped"
      ],
      "properties": {
       "source": {
        "type": "string"
       },
       "fetchedAt": {
        "type": "string",
        "format": "date"
       },
       "records": {
        "type": "integer",
        "minimum": 0
       },
       "pages": {
        "type": "integer",
        "minimum": 0
       },
       "candidates": {
        "type": "integer",
        "minimum": 0,
        "description": "공동주택 + 총세대수>0 인 번지·블록 수"
       },
       "projects": {
        "type": "integer",
        "minimum": 0
       },
       "located": {
        "type": "integer",
        "minimum": 0,
        "description": "지도에 올린 수(features 길이)"
       },
       "unlocated": {
        "type": "integer",
        "minimum": 0
       },
       "unlocatedNames": {
        "type": "array",
        "items": {
         "type": "string"
        }
       },
       "unlocatedList": {
        "type": "array",
        "items": {
         "type": "object",
         "required": [
          "name",
          "jibun",
          "status",
          "reason"
         ],
         "properties": {
          "name": {
           "type": "string"
          },
          "jibun": {
           "type": "string"
          },
          "status": {
           "type": "string"
          },
          "units": {
           "oneOf": [
            {
             "type": "integer"
            },
            {
             "type": "null"
            }
           ]
          },
          "approvedAt": {
           "oneOf": [
            {
             "type": "string",
             "format": "date"
            },
            {
             "type": "null"
            }
           ]
          },
          "reason": {
           "type": "string"
          }
         }
        }
       },
       "blockProjects": {
        "type": "integer",
        "minimum": 0,
        "description": "공공주택지구 블록 단위 허가 중 대장으로도 위치를 못 정한 수"
       },
       "blockList": {
        "type": "array",
        "items": {
         "type": "object",
         "required": [
          "name"
         ],
         "properties": {
          "name": {
           "type": "string"
          },
          "block": {
           "oneOf": [
            {
             "type": "string"
            },
            {
             "type": "null"
            }
           ]
          },
          "units": {
           "oneOf": [
            {
             "type": "integer"
            },
            {
             "type": "null"
            }
           ]
          },
          "status": {
           "type": "string"
          },
          "approvedAt": {
           "oneOf": [
            {
             "type": "string",
             "format": "date"
            },
            {
             "type": "null"
            }
           ]
          },
          "records": {
           "type": "integer"
          }
         }
        }
       },
       "ledger": {
        "type": "object",
        "required": [
         "used"
        ],
        "properties": {
         "used": {
          "type": "boolean"
         },
         "bjds": {
          "type": "integer"
         },
         "rows": {
          "type": "integer"
         },
         "blocksMatched": {
          "type": "integer"
         },
         "blocksAmbiguous": {
          "type": "integer"
         },
         "parcelsRecovered": {
          "type": "integer"
         },
         "completions": {
          "type": "integer"
         },
         "error": {
          "type": "string",
          "description": "건물대장 보강이 완전하지 않은 이유(이때 응답은 짧게만 캐시)"
         }
        }
       },
       "parcelErrors": {
        "type": "integer",
        "minimum": 0,
        "description": "필지 경계를 일시 오류로 못 받은 수(0이 아니면 응답은 짧게만 캐시)"
       },
       "truncated": {
        "type": "boolean",
        "description": "사업이 120곳을 넘어 세대수 큰 순으로 남김"
       },
       "skipped": {
        "type": "object",
        "required": [
         "notCandidate",
         "noPnu"
        ],
        "properties": {
         "notCandidate": {
          "type": "integer"
         },
         "noPnu": {
          "type": "integer"
         }
        }
       }
      }
     }
    }
   },
...
   "InfraResponse": {
    "type": "object",
    "required": [
     "type",
     "bjd",
     "asOf",
     "sources",
     "schools",
     "stops",
     "meta"
    ],
    "properties": {
     "type": {
      "const": "Infra"
     },
     "bjd": {
      "type": "string",
      "pattern": "^\\d{10}$"
     },
     "asOf": {
      "type": "string",
      "format": "date"
     },
     "sources": {
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/InfraSource"
      }
     },
     "schools": {
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/School"
      }
     },
     "stops": {
      "type": "array",
      "items": {
       "$ref": "#/components/schemas/Stop"
      }
     },
     "meta": {
      "type": "object",
      "required": [
       "centers",
       "schools",
       "stops"
      ],
      "properties": {
       "centers": {
        "type": "integer",
        "minimum": 0,
        "description": "인허가 단지 필지 중심 수(상한 30)"
       },
       "schoolsNational": {
        "type": "integer"
       },
       "schools": {
        "type": "integer"
       },
       "stopCalls": {
        "type": "integer"
       },
       "stops": {
        "type": "integer"
       },
       "schoolsError": {
        "type": "string"
       },
       "stopsError": {
        "type": "string"
       },
       "noBus": {
        "const": true,
        "description": "TAGO·서울시·OpenStreetMap 어디에도 정류장 자료가 없다. 교통 점검은 '자료 없음'"
       },
       "stopsSource": {
        "type": "string",
        "enum": [
         "tago",
         "seoul",
         "osm"
        ],
        "description": "정류장 출처: tago(국토교통부) 우선, 서울(시도 11)은 seoul(서울특별시 정류소정보조회), 그 밖에 TAGO 에 자료가 없는 지역(강릉 등)과 서울시 조회 실패 때는 osm(OpenStreetMap, ODbL)"
       },
       "seoulError": {
        "type": "string",
        "description": "서울시 정류소 조회가 실패했거나 일부만 받았다(시간당 상한·인증키 한도·연결 오류). 이때 응답은 s-maxage=60 이고 OpenStreetMap 으로 물러났을 수 있다"
       }
      }
     }
    }
   },
...
  }
 }
}